yaml: test and fix include of scalar with variables
Test including a string with variables, and setting those variables.
This requires some changes in t_yaml_env_do_include, to avoid
parsing the subfile in a stack-allocated yaml_data_t, so that the
pointer in yaml_variable_t do not become invalid.

Clearly, this is not ideal. The code in t_yaml_env_do_include was
perfectly sane, the issue is in the yaml_variable_t impl. A FIXME is
added about this, it really needs to be fixed some way or another.

Change-Id: Ic6d4dcecd93b684244f69e267ccf75345a041e64