zchk-iopy: fix calling iopy/z_run_tests.sh with another waf environment.
When `z_run_tests.sh` is called with `waf check` but the actual
environment on which `z_run_tests.sh` is different from waf, the call to
`waf list` will fail. This can happen if the PATH is modified on waf
execution.

The solution is to avoid using `waf list` in `z_run_tests.sh`, and just
check for the compiled files existence.

Change-Id: I652cf9cea56d31458d6568cf46a44af623d943d4