zchk-iopy: suppress python leaks on ASAN.
When zchk-iopy-dso is run in ASAN it will complain about leaks in
libpython.
This is actually normal as Python, by design, will not properly clean
all the objects tracked by the garbage collector at exit.

So, to avoid making the tests fail, we need to suppress the leaks
coming from libpython.

Change-Id: Id7bc6da9b417ef7571cf0a8136abbd030b8a264b