yaml: rework include syntax to generate valid YAML
The previous syntax was as followed:

!include <path>
  variables:
    <vars>
  <overrides>

this isn't great because it is not valid YAML: we have a tagged string,
followed by another data, where we should have a single data.

Instead, change the syntax to put the path in the tag:

!include:<path>
  variables:
    <vars
  <overrides>

This is properly parsed as a tagged object, making this more robust, and
removing specific code to handle parsing of the object.

This has significant changes in the code:

* For parsing, we now directly have the whole vars + overrides data when
parsing the tagged object, and no longer have to parse it separately.

* For packing, we need to reorder the steps a bit:
** First, pack the subfile.
** Then, generate the vars & overrides data (must be done after packing
the subfile, as this packing will use vars & overrides).
** Finally, packed the tagged object.

* The presentation is changed a bit as well. As the include tagged
object now contains variables & overrides, the document presentation is
no longer needed in the override IOP, and can be put in the include IOP.

Change-Id: Idb8327d9d4f7f960f9863cbf1f1f15df81ba0cfb