yaml: change syntax to set variables
Instead of setting variables with this syntax:

!include foo.yml
$(var1): 3
$(var2): 5

Use this syntax:

!include foo.yml
variables:
  var1: 3
  var2: 5

This removes all the specific code about parsing keys as variables,
simplifying the code greatly.

The only downside, is that the "variables" string is now reserved, and
it is thus not possible to override this string. This could be fixed
using an additional more specific syntax, but for the moment, is left as
a TODO.

Change-Id: I1e876923c8f7dd33db92f37b8da74a52f532949a