yaml: fix repacking for some shared scalar variables
An innocent template like this one:

id: $(id)
name: name-$(id)

!include:inner.yml
variables:
  id: 2

will get repacked into:

id: $(id)
name: name-$(id~1)

!include:inner.yml
variables:
  id: 2
  id~1: 2

Why? because the first variable is a uint scalar, while the second one
is a string. The comparison of the two scalars fails, so a new variable
is created to handle the difference.

To fix this, we need to relax our comparison, the same way the parsing
is handled. When parsing, if a variable used in a string is not a
string, it is converted to a string by taking its string representation
from the parsed document.

When repacking, we thus want to compare the string representations of
the deduced values, so that the above example can work.

This is done by making yaml_data_equals implement weak comparison.
Unfortunately, We have to duplicate a little bit the scalar -> string
behavior, as we cannot use yaml_pack_scalar, it is too complex and
handles too many things for our little case.

This is tested with multiple types of scalars. Also, a test shows how
using a non-canonical representation for a scalar will make the
variables mismatch on repacking. This could be fixed by using the
canonical version when parsing, but this is very unclear what would be
the desired behavior for a user:

template: <$(var)>

!include:inner.yml
variables:
  var: null

should we generate "template: <null>" or "template: <~>"? Imho the first
option makes much more sense, and it is what we do.

Change-Id: I0a3b950dcc7b3859bef7ec4dde3348487cbc67cb