qhash.c: align qhash/qmaps on biggest alignments
If a struct is supposed to be aligned on 16 bytes, using it inside a
qhash/qmap can lead to segfaults (in my case, compiling with
optimizations lead to the use of the MOVAPS instruction which expects
a 16-byte alignment). The 16-byte alignment is for example used with
long doubles.

So, the qhash/qmap allocators should use __BIGGEST_ALIGNMENT__ and not
8.

This __BIGGEST_ALIGNMENT__ was used in two other files, leading to some
duplication, so its definition (if the compiler does not provide it)
is moved in core-macros.h.

Change-Id: I4f5f8ca3da8b350636e7297c5aa1991a0e3c4a85