yaml: implement merge key handling on parsing
Handle merge key in an object when parsing. This merge key must be
placed first in the object, and can be either an object or a sequence.
All fields from this merge key are inlined in the object. If a key
already exists, the value is replaced.

This is useful to separate group of values even if they belong in the
same object.
For example, values can be put in separate files, which can then be
inserted through includes in a merge key:

<<:
  - !include:conf_foo.yml
  - !include:conf_bar.yml
key1: ...
etc

Change-Id: I6b8827ca4492f78df74594a4c3057fbd1ba81031