yaml: allow relative includes in root directory
Any includes that was not in the same directory was previously
forbidden. This ensures that all the document is contained in the same
directory, and a "dump" will not create file anywhere.

However, this prevents legitimate use-cases, where all the document is
contained in a directory, but includes go to siblings directories:

conf/
  cfg.yml
  templates/
    foo.yml
  services/
    bar.yml <-- does includes from ../templates/foo.yml

To allow this, we now keep track of the "root" directory, ie the
directory of the entry file. As long as includes stay in this directory,
they are allowed.

Change-Id: Iae00de52a16043769cccfe72b7fec8fd5a775a8a