yaml: Save merge key details in presentation
Implement presentation for merge keys. This is similar to overrides,
but limited to a single level. The implementation is slightly complex
due to how an object with a merge key works. In specific, all keys
specified after the merge key still behave as if coming from a merge
key (otherwise, key collisions with keys coming from the merge key would
lead to errors).

Change-Id: Ic8bf73e4217e91b0b79b0ee04c2d77b0d911d1eb