iopc: fix IOP typescript generation for recursive parent dependencies.
Since 5b77a3747af2716ac66979e005e43bb0de650f6a, we consider all the
dependencies recursively of a class instead of just its immediate
parent.

This working fine when generating the C structs associated to the class.
But when generating typescript packages, this can be an issue because
the dependency is not used directly.
This can generate some import that are not actually used in the
typescript package:
TS6133: 'test__test' is declared but its value is never read.

The solution is to simply not consider the dependencies recursively when
generating the typescript packages. This way, only the immediate
dependencies are used, and we are sure that all import are used.

Change-Id: I9e3cf13ee410566d2b99ebe99d26de7dfd953562
Refs: #71098