ichannel: Various fixes in ic_mark_connected.
This function had two issues:

* an obvious one, which is that ic_disconnect can delete "ic", and thus
  accessing it just after can be invalid.

* a less obvious one: calling ic_mark_disconnected from ic_read (which
  happens in probably more than 90% of the time because it is usually
  called from a RPC implementation callback) was invalid because in the
  case the ichannel is spawned and autodel, it was deleted by
  ic_disconnect, and thus when getting back in ic_read there were invalid
  reads, which could lead to crashes.

To fix that, just set is_closing to true before calling ic_disconnect,
which prevents the ichannel deletion. When getting back to ic_read, the
error will be properly catched (because ic->elh will be NULL), and the
ichannel properly destroyed at the end of ic_event.

Change-Id: I5481a0a560d51d4dc284cc4d04906b5e763a471e
Refs: #71460