sctp: add function sctp_peer_addr_state_is_active.
Since Centos/RHEL version >= 6.4 (kernel version 2.6.32-358), some
backward incompatible modifications were done in the SCTP kernel
headers. The SCTP_ACTIVE value is now equal to 2 when it was previously
equal to 1. In addition this change isn't reported in the REHL headers,
so at compile-time we think it's 1 for all REHL 6.

We have no other solution than to detect the right value at runtime
depending on the kernel version.

The new function sctp_peer_addr_state_is_active() allows to correctly
check if the state of a peer is active or not.

Note: this function should be removed when Centos/RHEL 6 support will be
dropped.

Change-Id: If145d590a807a42a6482b06831cbf7ed41e9025b
Refs: #68528