fix compilation with GCC < 4.5 because of __builtin_unreachable
a0d64eb8c4bc1436df6e37b9d22ec7f4d8594f49 introduced the usage of
a __builtin_unreachable which is only available since GCC 4.5, so let's
replace it by an assert(false) when an older GCC is used. It's clearly
not strictly equivalent but it will be an acceptable replacement on the
old branches.

Change-Id: I9d41a8fb69940316a10b9a640801af3fb6938385