iop: handle struct -> class backward compat on binary unpacking
Changing a struct to a class is not handled when unpacking in binary.
However, it is not that hard to implement.

A class is encoded as is:
 * len of the whole class
 * class_id of the real class with tag 0
 * pack of the class fields (as if the class was a struct)
 * for every parent, if non empty:
    * pack class_id of the parent, with tag 0
    * pack of the parent fields

ie:

| LEN | class_id(A) | fields(A)... | class_id(B) | ...

A struct is encoded as is:
 * len of the whole struct
 * pack of every field

ie:

| LEN | fields... |

The two are almost equivalent if:
 * the class_id tag is ignored
 * the class does not have parents. Strictly speaking, having parents
   with only optional fields would be backward compatible, but that was
   not tested).

To handle the compat, it is thus enough to:
 * detect if the class_id is missing.
 * make sure the class has no parents (as said above, that may not be
   necessary...)

The backward compat checks are updated as well. An error is returned for
JSON unpacking when a struct is transformed into a child class. This is
actually backward compatible, but the rest of the backward compat checks
do not handle it correctly, and testing it leads to errors. This is left
as a future fix.

Refs: #71836
Change-Id: I5a35b7ef180a2bff47a04a00fcc0e62975945e22