waf: fix sorting list of Node in get_old_gen_files() for python3.
Python 3 does no longer support sorting list of objects of type that
does not implement the '<' operator natively.

Instead, we should use the `key` argument to the sorted function to
provide the comparison key to be used on sort.

In the case of a list of Nodes, we use their absolute path.

Change-Id: I3ead17317497bd55d06176716d051da557e79067