iop-yaml: handle bytes type on JSON subfiles => yaml conversion
When converting JSON subfiles includes into YAML presentation, we want
to detect when the include are related to a raw document, so that
"includeraw" is used instead of "include". This was done by detecting
that the include is for a string field, but this is not enough. For
example, the data type can also be used (for example, with
core.TlsCertAndKey).

So the code is generalized a bit, and any include on a string/data/xml
field will generate an includeraw.

Closes: #72323
Change-Id: I0d5f288a95742cc9028f185d959804f61a8e99c6