yaml: add binary handling
Previously, the handling of binary data was done not in the YAML parser,
but when deserializing as IOP objects. If a field was of type byte,
the data was expected to be base64 encoded, but the YAML AST only
handled strings.

This causes an issue however, due to the !includeraw feature. It is
possible to include a raw file (which can be binary) into a field. The
contents of the file is then imported raw, which causes issues if the
field is of IOP type bytes:

 # Inline version
 data: SGk= # Must be in base64

 # include version
 data: !includeraw:file # file directly contains the binary

When repacking, the IOP => AST generator cannot know (and should not
know) if a field is in an include or not. However:
* if it isn't, it should base64 encode the data
* if it is, it should not encode the data

To fix this, we need to move the binary handling into the YAML AST
directly:

* A new type of scalar, YAML_SCALAR_BYTES, is added.
* This type can be declared using the "!binary" tag, and expects a
  base64 encoded value.
* The base64 logic is removed from iop/yaml.blk: the IOP bytes data maps
  to the yaml bytes data.

So the correct way to write the above example is now:

data: !binary SGk=

This value is decoded by the parser, and properly read into IOP. When
reserialized, a scalar of type BYTES is generated, which can be properly
handled in both the inline and the includeraw context (the includeraw
handling will come in the next patch).

This is clearly not backward compatible. However, there isn't any YAML
files in production for the moment, so we can do this change.

Refs: #72325
Change-Id: Icd1eb5f4f9923a30d8c8de128a5bf8538e2e76fb