asn1: avoid confusion between OPTIONAL OPEN TYPE field and EOC
This patch avoids the confusion that could happen between OPTIONAL OPEN
TYPE field and EOC (00 00). This case happens when a SEQUENCE is encoded
using indefinite length (in that case, an EOC is appended after the last
field of the SEQUENCE).

Previously, any tag was expected for an open type or an untagged choice,
but still we should skip the tag 0 in order to avoid confusion with EOC.

This not a very usual behavior but it was observed in some TCAP traces.
Non regression tests will be added to the TCAP library.

Change-Id: I6ed52d40206bec9b11f61411d14cdfc9d88a92c6