coding-rules: update 4 existing coding rules
Our coding rules have several idiosyncracies that:
* are surprising to newcomers, and takes time and energy on code review
  to enforce.
* are unknown to formatting tools, making it basically impossible to
  auto-format files.

After discussion, a general consensus has emerged to remove several of
those rules:

* No space on function call, expect for asserts (Personally, one rule
  that never made sense to me).

 fn(1);
 assert (false); /* previously */
 assert(false); /* now */

* Double space after a macro name.

 #define FOO  1 /* previously */
 #define FOO 1 /* now */

* newline between else and if.

 /* previously */
 if (a) {
     b();
 } else
 if (c) {
     d();
 }

 /* now */
 if (a) {
     b();
 } else if (c) {
     d();
 }

* Operand alignment on multiline statement.

 /* previously */
 if (a
 ||  b)
 {
 }

 /* now */
 if (a ||
     b)
 {
 }

The coding rules doc is updated to fit those new rules. Obviously, the
existing files do not match those changes. The rule is to update code
to fit the new coding rules when a patch is done in existing code. Or
maybe someone will sed the codebase to reflect the new rules.

The goal is to also provide soon a standardized clang-format
configuration file that matches our rules. Unfortunately, that file
will only work with clang-format 12 at least, as it requires this patch:
https://reviews.llvm.org/D85304

Change-Id: I212e13402431a7fe66a5bd898b1769c9043ac60f