yaml: fix invalid spans on obj/seq ending with includes
Every parsed object has a span that is used when an error is triggered
on the object. When parsing an object or a sequence, the end of the span
is retrieved from the last parsed element, so that the span can be a bit
more precise:

a: - 1
   - 2
b: 3

The span for the seq is: from line 1, char 4 to line 2, char 7.

The issue is that if the last element is an include, its span is
relative to the included file, and not the current file. This makes the
span invalid and causes a crash if an error is triggered on the object.
To fix this, we revert the span optimization, and get the end of the
span from the current env. This generate a span a bit worse, but we
ensure it is valid.

a: - 1
   - 2
b: 3

The span for the seq is now: from line 1, char 4 to line 3, char 1.

An easy way to test this is to make a sequence override, which is
invalid.

Closes: #73983
Change-Id: I57d1032b062ac43ad01f96f2c0b23cb909818a61