yaml: test use of variables in previously forbidden contexts
Due to how we used to handle variables, there was a lot of restrictions
on where variables could be used: not in overrides, not in variables
definition, and the use in merge keys was buggy as well (see #73984).

The handling of variables have been reworked in the previous commit, and
this removes all those restrictions. So test those, and test as well the
use of variables with a merge key.

Closes: #73984
Change-Id: I05f956b10b9479fc8c1642e39f8214be1fefbda0