hash: introduce a case insensitive jenkins hash and use it in qhash
The previous implementation of the case insensitive ASCII qhash was
quite sub-optimal and not fit for the lib-common quality expectations.
The problematic function was qhash_lstr_ascii_ihash which:
    - performed a memory allocation on the t_stack;
    - duplicated the input string to hash;
    - walked twice over the input string to lower it then hash it.

There is probably an optimal way to implement a case insensitive hash
table, but for now we will just replace the qhash_lstr_ascii_ihash
implementation by something acceptable in the same spirit.

So first, we introduce the jenkins_hash_ascii_lower() function, which
implements the Jenkins hash but by applying a tolower() on each
character of the input data on the fly. Here we use the Jenkins hash
function instead of the MurmurHash for simplicity. Jenkins walks over
the input stream one byte at a time unlike the MurmurHash.

Here are the results of some benchmarks using the ztst-container
utility (insertion of 100K words of random lengths repeated 1k times):

Str len | standard QH | case ins. QH | old case ins. QH | new vs old
1 -  10 |    22.061ms |     20.944ms |         34.115ms |  +42%
1 -  30 |    22.657ms |     22.657ms |         39.913ms |  +36%
1 - 100 |    29.507ms |     42.318ms |         55.293ms |  +25%

Change-Id: I29d73cce961cabe1a9b65fcc9f3f4c4143f9c824