iop_field_path: add new iop_obj_get_field_type function
The iop field path only uses the IOP struct description when compiled.
This is an issue in some use-cases, when the path has been created from
a specific value, and goes through IOP classes.

For example, in the iop field path used for IOP json include paths,
we can find:

arrayOfClasses[2].fieldOfOneCfgChild

When compiling this path, the compilation will fail, as
"fieldOfOneCfgChild" is not a field of the base class.

To handle those usecases, an optional IOP value can be given to
t_iop_compile_field_path. If provided, and an IOP class is found,
the path is applied on the value, and the real type is retrieved from
the value.
Of course, this isn't compatible with wildcards (or at least, it is far
from trivial). So if a wildcard is found, the value is no longer used.

To avoid mixing this behavior with the existing one, this behavior is
only exposed through a new function, iop_obj_get_field_type, which
encapsulate the compile and the get_type, so that it is not possible to
call for example one with a different value that the one used for the
compile.

Change-Id: Iefae89841a9764f54c787d388c18dacfc9baf7d7