iop-yaml: fix edge cases in json included subfiles conversion
Two bugs in the conversion of iop json included subfiles are fixed:

* the prefix test was not properly done, and for two fields following
each other, with one being the prefix of the other, an include was
generated:

foo
foobar

this example generated ".bar!" inside the ".foo" include.

* the subfiles are sorted before conversion. This is required because
it is not guaranteed that the nested included are listed in order in the
json included subfiles.

Change-Id: I74eb70840342c6a3d5913591f59f9828f45cf3b4