qps: reestablish snap_syn attribute
Despite the absence of use in the lib-common, this attribute was directly
used at user-level.

Also rename the attribute as "snapshot_syn" and add a getter so we can
enforce proper encapsulation.

We should consider "hiding" qps_t structure in a private header and only
expose a handle to avoid this kind of encapsulation violation.

Change-Id: I8309c7e94cd16697a5fbf770796b04fa16e902b2