yaml: allow override of null values by any type of data
The 'null' value in yaml is special, as it is somewhat upgradable to
anything: it is valid to use 'null' for an array or object, and it
indicates an empty array/object.

This means that overriding a null value by an array or object should be
valid, but it was for the moment rejected. This is now fixed by
considering a 'null' value upgradable into any type when overridden
(including scalars).

Refs: #75113
Change-Id: I71f74215cf996ef75970bad3d479718507fc5201