datetime: Support timestamps in milliseconds in time_parse.
time_parse takes as argument a string that contains a date, and tries to
convert it in a UNIX timestamp expressed in seconds.

It supports various input formats, including UNIX timestamps, of course.
It assumed that the incoming timestamps were expressed in seconds and did
not support other units. However, the millisecond timestamps are quite
common, so it is interesting to support them, which is the purpose of
this commit.

The code now arbitrarily assumes that if it receives a timestamp greater
than 30000000000 (which is in 2920 if expressed in seconds) are actually
millisecond timestamps and are converted in seconds.

Refs: #76321
Change-Id: I72c5014ad10e4f2206df9e7b49f19ead1e20bbe6