waf: allow additions to local vimrc file
The local vimrc file is written by code in backend.py. However, we want
to be able to add lines in it from other files not related to
backend.py.

In order to do so, allow adding callbacks in 'ctx.vimrc_additions',
and iterate on it when generate the local vimrc file.

Change-Id: I205bcc48e26f497822d4ab999b0ad3f56154270b