core/errors: user context printers for .debug file
Add the capability of registering user context printers to enrich the
information already written to .debug files in case of crash.

This patch adds a new macro "debug_stack_scope()" that allows to declare a
couple (data, printer_callback) for the current scope. In case of crash
inside the scope, the data will be printed into the .debug file by the
dedicated printer.

Change-Id: I4776ae1d632c3db11fdf4511f11031e88cf5d0cb