iopc: Fix has_comments in iopc-lang-c.blk.
This patch follows 5763feb83813, that was made 5 years ago and was buggy.

With what was done, the iop_help_t of an object that had only an example
was not generated. And because of that, some examples written in our iop
files were not present in the .iop.c generated file (as pointed by the
update of tstdox.ref.c in this patch).

Fix has_comments, that now uses the same way as iopc_dump_help do detect
if there will be something to dump or not.

Refs: #77415
Change-Id: I883736f8e4dd052f75881428c20eb66b9ffa8a9c