lib-common: fix iopc runtime compile error with GCC 10.
This commit is a partial cherry-pick of 9fabe2747a9f.

When running iopc with GCC 10, it was emitting some cast-function-type
errors in iop_hash.
This is fixed in 2019.1 with 9fabe2747a9f, but unfortunately, we cannot
cherry-pick the whole patch in 2016.4.

The solution is to only apply the modifications relative to the hash
functions.

Change-Id: I15669c4bb1b3df5a6cf21c7328b30a72c91e43e9