lib-common/hash: use ssize_t for len of hash functions.
The hash functions imported from platform with 99815f9b113, needs to
have the same arguments as the iop hash functions.

With 6c709c9cf64, we now use ssize_t for the len argument for the iop
hash functions.
So, we need to do the same modifications as 9fabe2747a9f, and use
ssize_t for the len argument of the functions imported with 99815f9b113.

Change-Id: I1bae9a7c8d0ca91802c8c900f7aeb0a6ebd68770