build: stop to force Dwarf 2 in debug profiles
Back in 2015 we forced debug profiles to stick to Dwarf 2 because of
some incompatibilities with gdb and valgrind (see
45a8eb6ec6d5b74211c8e7777e0e99522d9976b5). It's about time to move
forward and stop using Dwarf 2.

If gcc 11 (using Dwarf 5) is released too soon we may have to force
Dwarf 4, but it's preferable to use the system defaults while everything
works correctly.

Change-Id: I4f8f33fb6793b05785029698d5b71ee6d5457ffb