lib-common/qlzo: fix miscompilation of lzo_copy_backptr() with GCC 10.2.
lzo_copy_backptr() used a micro optimization for copying blocks of 8
bytes at a time.
Unfortunately, due to an optimization bug in GCC 10.2
(https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100410), the blocks were
copied 16 bytes at a time, which broke the behaviour.

The solution is to use memcpy() instead of the micro optimization.

After doing some benchmarks, The memcpy() solution is a little slower
than the micro optimization (1.3%), but the benchmark was testing
exclusively lzo_copy_backptr() and not the whole LZO decompress
algorithm. So, the difference should be acceptable.

Change-Id: Ie1cd0204358ca86805771b32455758f502309e65