lib-common/bit-buf: fix sb buf alignment with GCC 10.
same patch as e9758b71036a04a5240367f94785270a0c78268e

With optimizations, GCC 10 does not align the buffer of SB(sb, ) on 8
bytes alignment. This triggered an assert in bb_init_full() on zchk asn1
per test.

The solution is to use a t_SB() which uses the t_scope and make
sure the buffer is align on 8 bytes.

Change-Id: Ib60ba8645c470a59e4e4447686294497e4ad98f2