lib-common/bit-buf: fix sb buf alignment with GCC 10.
With optimizations, GCC 10 does not align the buffer of SB(sb2, 42) on 8
bytes alignment. This triggered an assert in bb_init_full() on zchk
bit_buf.sb test.

The solution is to use a t_SB(sb2, 42) which uses the t_scope and make
sure the buffer is align on 8 bytes.

Change-Id: If561dc41dacce80c7efecd3cf87afc448ef4eee1