asn1/macros: check the struct prefix when registering a field
Add a "check variable" that is expected to exist in the context where each
field is registered. If a field is registered with the wrong "st_pfx", the
"check variable" will mismatch and the compilation will fail.

This patch is a miniature version of a biggest rework done in master that
makes this check inherent to the macro (the "check variable" is the
description variable), so this patch can be dropped when merged into
master.

Change-Id: I5ebd92ea05687e7c84f77afe9b77321ed44f968d