lib-common/rpc: fix invalid pointer to compound literal on local ic.
When using ic_build_query(), the arguments of the ic query are created
using a pointer to a compound literal and stored in the message.
With a local ic, the arguments are not deeply copied and only the
pointer is stored in the message.

According to the C99 standard and with GCC9+ with optimizations, a
compound literal is only valid on the block scope.
Since the message created by ic_build_query() has a longer lifetime than
the arguments, a compound literal, keeping a reference onto it is not
valid in case of a local ic.

That is why ic_build_query() is a dangerous API and cannot be used for a
local ic.
Fortunately, this macro is not used outside the ic_query*() macros can
easily be replaced with ic_build_query_p() with arguments with a longer
lifetime than the message.

Since ic_build_query() can't be used safely, it is best to remove it.

Change-Id: Icac8eb423d7fbdb916e01cae75bff0c053c07645