iop: add forceFieldName IOP attribute.
In the scope of the 5G support, we want to be able to translate the
structure described in the 3GPP standard with openAPI into IOP
structures.

Some of them do not respect the camel case notation, which is not
allowed in IOP format. The forceFieldName attribute is used to provide
the name exactly as it should be encoded in the JSON format.

Closes: #77869
Change-Id: I5ba6a76a0d6f58fa7370dd57d23067a878915968