asn1/aper: prepare fragmentation support in 'aper_read_ulen()'
Prepare for fragmentation support: peek the first byte instead of
reading it so the function can stop without touching the bit-stream if a
fragmented object is detected.

Also simplify the function a bit: get rid of the anonymous union.
Purists will appreciate the gain in portability. I just expect the
function to be slightly more readable.

Refs: #78540
Change-Id: I989a0c8e81db9432a2cb9fc2d210a892f960b074