asn1/aper: rewrite low-level tests for length encoder/decoder
Use Z helpers instead of anonymous struct arrays: in case of failure,
the diagnostic can be way more straight forward and can even be done
without launching the test in gdb.

With the struct arrays, the developer first has to check the index of
the test case that failed to understand what went wrong. With a
Z_HELPER, the developer gets a proper error stack.

Refs: #78540
Change-Id: Ieb6fc013c4dc26b1e79923d71ace3da8177c180f