asn1/aper: support decoding of fragmented SEQUENCE OF
In ASN.1 PER/APER, when the length of a SEQUENCE OF is unconstrained (or
within the extension) and its values is 16k or more, we have to use
fragmented encoding.

This patch uses the existing helpers ('aper_len_encoding_ctx_t') to
implement fragmented SEQUENCE OF decoding.

Note: the code could look better and simpler. Expect one or two
improvement patches.

Refs: #78540
Change-Id: Ie096b3f954e500e32125044a5f9e89531f744d68