asn1/aper: rework fragmented decoding helpers
Adapt the helpers for decoding fragmented data so it can be used this
way:

    aper_decode_len_extension_bit(bs, info, &ctx);
    do {
        aper_decode_len(bs, ctx);
        /* Decode <ctx.to_decode> element(s). */
    } while (ctx.more_fragment_to_read);

Use the helpers to unify and simplify the decoders for both OCTET STRING
and SEQUENCE OF.

OCTET STRING has no special decoder for fragmented content anymore. Less
code == fewer bugs.

The last data type that doesn't support fragmentation yet is BIT STRING.

Refs: #78540
Change-Id: Iae610a02a1e7538ec378823f0c2a04bb4ad62f2d