asn1/aper: prepare tests for fragmentation support for BIT STRING
When the length an unconstrained BIT STRING is 16k or more, it has to be
encoded by fragments. This patch adds a test that only checks that we
properly emit an error when trying to encode a BIT STRING that should be
fragmented. Next patches will add actual support.

Also add a BIT STRING unit test, initially intended to unveil a bug that
didn't exist, but one more test won't hurt.

Refs: #78540
Change-Id: Id080412b2d228bcf5317deff84df6a0228004486