asn1/aper: simplify decoding for BIT STRING
The function 't_aper_decode_bstring()' has a 'copy' parameter but the
only caller of the function (if we exclude unit tests) never actually
used it because it copied the result into a bit buffer anyway.

This patch removes the 'copy' parameter and makes the function write
directly in a bit buffer. It is going to make the decoding of fragmented
bit strings way easier.

Remark about the tests: the removal of the 'copy' parameter allow us to
remove one layer of testing helper. We also switch to automatic cleanup
for the bit buffers we use there: it will prevent ASAN from generating
misleading leak reports when a test fails.

Refs: #78540
Change-Id: I40c6c5092fd28be22c24f2556dfb5d4204edd0fb