asn1/aper: add a test for SEQUENCE OF written to arrays of pointers
After the code changes we made in SEQUENCE OF encoding/decoding, the
allocation of array of pointers stopped relying on common ASN.1 helper
'asn1_alloc_seq_of()', but we didn't have any APER unit test for this
case, leading the allocation to be untested.

This patch adds a specific test that checks that the APER library
properly allocates arrays of pointers so we don't have this blind spot
in or tests.

Refs: #78540
Change-Id: Iba59df838419baefa51e9e07328b1b506f92c7cc