asn1/aper: add tests for "indefinite length" case
This patch adds tests for the indefinite length case (see below) without
trying to fix anything, as the specification is quite unclear about this
case.

The specification we use as a reference for ASN.1 PER is the ITU-T
X.691. This document has an enigmatic sentence to explain how a
constrained length should be encoded when <MAX_LENGTH> - <MIN_LENGTH> is
>= 64k:

X.691 §11.5.7.4:
> [in this case], the value ("n" – "lb") shall be encoded as a
> non-negative-binary-integer in a bit-field (octet-aligned in the
> ALIGNED variant) with the minimum number of octets as specified in
> 11.3, and the number of octets "len" used in the encoding is used by
> other clauses that reference this subclause to specify an encoding of
> the length.

The end of the paragraph is quite cryptic when it comes to the "len"
used in the encoding but we have to understand that the way the "len" is
encoded is the same as the one in another clause that references the
clause 11.5.7.4 and this clause seems to be §13.2.6 a).

Our implementation doesn't seem to respect the specification but the
point of this patch isn't to fix this, but just to freeze the current
behavior before we actually decide to fix it (it doesn't concern common
cases).

Refs: #78540
Change-Id: I943902e6abf5dfefccdfa57af63c90f80a190f27